Audit selection agnostic components
Split selection spec-conformance and property tests into per-component modules, document the Epic #221 audit, sync selection specs with public event surfaces, and fix SegmentGroup controlled value sanitization for disabled items.

Entire-Checkpoint: dde1b6540daf
