+++ crates/ars-components/tests/spec_conformance/selection/combobox.rs
@@ -1,0 +1,54 @@
+use ars_components::selection::combobox;
+
+use super::{Key, assert_anatomy};
+
+#[test]
+fn combobox_anatomy_matches_spec() {
+    assert_anatomy(
+        "combobox",
+        &[
+            (combobox::Part::Root, "root"),
+            (combobox::Part::Label, "label"),
+            (combobox::Part::Control, "control"),
+            (combobox::Part::Input, "input"),
+            (combobox::Part::Trigger, "trigger"),
+            (combobox::Part::ClearTrigger, "clear-trigger"),
+            (combobox::Part::Positioner, "positioner"),
+            (combobox::Part::Content, "content"),
+            (
+                combobox::Part::ItemGroup {
+                    key: Key::default(),
+                },
+                "item-group",
+            ),
+            (
+                combobox::Part::ItemGroupLabel {
+                    key: Key::default(),
+                },
+                "item-group-label",
+            ),
+            (
+                combobox::Part::Item {
+                    key: Key::default(),
+                },
+                "item",
+            ),
+            (
+                combobox::Part::ItemText {
+                    key: Key::default(),
+                },
+                "item-text",
+            ),
+            (
+                combobox::Part::ItemIndicator {
+                    key: Key::default(),
+                },
+                "item-indicator",
+            ),
+            (combobox::Part::Empty, "empty"),
+            (combobox::Part::Description, "description"),
+            (combobox::Part::ErrorMessage, "error-message"),
+            (combobox::Part::LiveRegion, "live-region"),
+        ],
+    );
+}
