+++ crates/ars-components/tests/spec_conformance/selection/listbox.rs
@@ -1,0 +1,48 @@
+use ars_components::selection::listbox;
+
+use super::{Key, assert_anatomy};
+
+#[test]
+fn listbox_anatomy_matches_spec() {
+    assert_anatomy(
+        "listbox",
+        &[
+            (listbox::Part::Root, "root"),
+            (listbox::Part::Label, "label"),
+            (listbox::Part::Content, "content"),
+            (
+                listbox::Part::ItemGroup {
+                    key: Key::default(),
+                },
+                "item-group",
+            ),
+            (
+                listbox::Part::ItemGroupLabel {
+                    key: Key::default(),
+                },
+                "item-group-label",
+            ),
+            (
+                listbox::Part::Item {
+                    key: Key::default(),
+                },
+                "item",
+            ),
+            (
+                listbox::Part::ItemText {
+                    key: Key::default(),
+                },
+                "item-text",
+            ),
+            (
+                listbox::Part::ItemIndicator {
+                    key: Key::default(),
+                },
+                "item-indicator",
+            ),
+            (listbox::Part::Description, "description"),
+            (listbox::Part::ErrorMessage, "error-message"),
+            (listbox::Part::LoadingSentinel, "loading-sentinel"),
+        ],
+    );
+}
