+++ crates/ars-components/tests/spec_conformance/selection/menu.rs
@@ -1,0 +1,91 @@
+use ars_components::selection::menu;
+
+use super::{Key, assert_anatomy};
+
+#[test]
+fn menu_anatomy_matches_spec() {
+    assert_anatomy(
+        "menu",
+        &[
+            (menu::Part::Root, "root"),
+            (menu::Part::Trigger, "trigger"),
+            (menu::Part::Positioner, "positioner"),
+            (menu::Part::Arrow, "arrow"),
+            (menu::Part::Content, "content"),
+            (
+                menu::Part::ItemGroup {
+                    key: Key::default(),
+                },
+                "item-group",
+            ),
+            (
+                menu::Part::ItemGroupLabel {
+                    key: Key::default(),
+                },
+                "item-group-label",
+            ),
+            (
+                menu::Part::Item {
+                    key: Key::default(),
+                },
+                "item",
+            ),
+            (
+                menu::Part::ItemText {
+                    key: Key::default(),
+                },
+                "item-text",
+            ),
+            (
+                menu::Part::ItemIndicator {
+                    key: Key::default(),
+                },
+                "item-indicator",
+            ),
+            (menu::Part::Separator, "separator"),
+            (
+                menu::Part::CheckboxItem {
+                    key: Key::default(),
+                },
+                "checkbox-item",
+            ),
+            (
+                menu::Part::RadioGroup {
+                    group: Key::default(),
+                },
+                "radio-group",
+            ),
+            (
+                menu::Part::RadioItem {
+                    key: Key::default(),
+                    group: Key::default(),
+                },
+                "radio-item",
+            ),
+            (
+                menu::Part::SubTrigger {
+                    key: Key::default(),
+                },
+                "sub-trigger",
+            ),
+            (
+                menu::Part::SubPositioner {
+                    key: Key::default(),
+                },
+                "sub-positioner",
+            ),
+            (
+                menu::Part::SubContent {
+                    key: Key::default(),
+                },
+                "sub-content",
+            ),
+            (
+                menu::Part::Shortcut {
+                    key: Key::default(),
+                },
+                "shortcut",
+            ),
+        ],
+    );
+}
