+++ crates/ars-components/tests/spec_conformance/selection/select.rs
@@ -1,0 +1,55 @@
+use ars_components::selection::select;
+
+use super::{Key, assert_anatomy};
+
+#[test]
+fn select_anatomy_matches_spec() {
+    assert_anatomy(
+        "select",
+        &[
+            (select::Part::Root, "root"),
+            (select::Part::Label, "label"),
+            (select::Part::Control, "control"),
+            (select::Part::Trigger, "trigger"),
+            (select::Part::ValueText, "value-text"),
+            (select::Part::Indicator, "indicator"),
+            (select::Part::ClearTrigger, "clear-trigger"),
+            (select::Part::Positioner, "positioner"),
+            (select::Part::Content, "content"),
+            (
+                select::Part::ItemGroup {
+                    key: Key::default(),
+                },
+                "item-group",
+            ),
+            (
+                select::Part::ItemGroupLabel {
+                    key: Key::default(),
+                },
+                "item-group-label",
+            ),
+            (
+                select::Part::Item {
+                    key: Key::default(),
+                },
+                "item",
+            ),
+            (
+                select::Part::ItemText {
+                    key: Key::default(),
+                },
+                "item-text",
+            ),
+            (
+                select::Part::ItemIndicator {
+                    key: Key::default(),
+                },
+                "item-indicator",
+            ),
+            (select::Part::HiddenInput, "hidden-input"),
+            (select::Part::Description, "description"),
+            (select::Part::ErrorMessage, "error-message"),
+            (select::Part::EmptyState, "empty-state"),
+        ],
+    );
+}
